hrms_backend/requirements.txt
Django
django
djangorestframework
django-rest-authtoken
django-cors-headers
django-rest-auth

whitenoise
dj-database-url
gunicorn
python-dotenv
django-ckeditor
Pillow

psycopg2
stripe
django-cities-light
channels
django-storages
azure-core
azure-storage-blob
beautifulsoup4
selenium
django-ipware
geopy
channels
pandas
dj-database-url
psycopg2